@@ -3,14 +3,12 @@ package com.advancedpwr.test;
 import java.util.ArrayList;
 import java.util.List;
 
-import junit.framework.TestCase;
-
 /**
  * @author Matthew Avery, mavery@advancedpwr.com
  * Created: Mar 14, 2016
  *
  */
-public abstract class BaseTestCase extends TestCase
+public abstract class BaseTestCase //extends TestCase
 {
 
 	protected List<String> fieldCalls;
@@ -40,17 +38,17 @@ public abstract class BaseTestCase extends TestCase
 		getCalls().add( methodName + " : " + inArgument );
 	}
 
-	public void failCall()
-	{
-		Throwable throwable = new Throwable();
-		String methodName = throwable.getStackTrace()[1].getMethodName();
-		fail( "should not have called " + methodName );
-	}
-	
-	public void assertCalls( String calls )
-	{
-		assertEquals( calls, getCalls().toString() );
-	}
+//	public void failCall()
+//	{
+//		Throwable throwable = new Throwable();
+//		String methodName = throwable.getStackTrace()[1].getMethodName();
+//		fail( "should not have called " + methodName );
+//	}
+//	
+//	public void assertCalls( String calls )
+//	{
+//		assertEquals( calls, getCalls().toString() );
+//	}
 
 	public void clearCalls()
 	{
