@@ -15,88 +15,73 @@
  */
 package com.advancedpwr.record;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+
 import java.io.PrintWriter;
 import java.io.StringWriter;
 import java.util.Set;
 
-import junit.framework.TestCase;
+import org.junit.jupiter.api.Test;
 
-public class ClassWriterTest extends TestCase
-{
+public class ClassWriterTest {
 	ClassWriter writer;
 	StringWriter out;
 
-	@Override
-	protected void setUp() throws Exception
-	{
-		writer = new ClassWriter()
-		{
+	public ClassWriterTest() {
+		writer = new ClassWriter() {
 
-			protected Set<Class> classes()
-			{
+			protected Set<Class> classes() {
 				return null;
 			}
 
-			protected ClassDescriptor createDefaultDescriptor()
-			{
+			protected ClassDescriptor createDefaultDescriptor() {
 				return null;
 			}
 
-			protected void writeObjectBuilderMethod()
-			{
-			}};
+			protected void writeObjectBuilderMethod() {}
+		};
 		out = new StringWriter();
-		writer.setWriter( new PrintWriter( out ) );
+		writer.setWriter(new PrintWriter(out));
 	}
-	
-	public void testWriteLine()
-	{
-		writer.writeLine( "package com.advancedpwr.foo" );
-		assertResult( "package com.advancedpwr.foo;\n" ); 
+
+	@Test
+	public void testWriteLine() {
+		writer.writeLine("package com.advancedpwr.foo");
+		assertResult("package com.advancedpwr.foo;\n");
 	}
-	
-	public void assertResult( String inString )
-	{
-		assertEquals( inString, out.toString().replaceAll( "\r\n", "\n" ) );
+
+	public void assertResult(String inString) {
+		assertEquals(inString, out.toString().replaceAll("\r\n", "\n"));
 	}
-	public void testWrite()
-	{
-		writer.write( "if ( foo == null )" );
-		assertResult( "if ( foo == null )\n");
+
+	@Test
+	public void testWrite() {
+		writer.write("if ( foo == null )");
+		assertResult("if ( foo == null )\n");
 	}
-	
-	public void testOpenCloseBrace()
-	{
+
+	@Test
+	public void testOpenCloseBrace() {
 		writer.openBrace();
-		writer.write( "if ( foo == null )" );
+		writer.write("if ( foo == null )");
 		writer.openBrace();
-		writer.writeLine( "bar" );
+		writer.writeLine("bar");
 		writer.closeBrace();
 		writer.closeBrace();
-		
-		assertResult( "{\n" +
-				"\tif ( foo == null )\n" +
-				"\t{\n" +
-				"\t\tbar;\n" +
-				"\t}\n" +
-				"}\n");
-		
-		try
-		{
+
+		assertResult("{\n" + "\tif ( foo == null )\n" + "\t{\n" + "\t\tbar;\n" + "\t}\n" + "}\n");
+
+		assertThrows(ClassWriterException.class, () -> {
 			writer.closeBrace();
-			fail( "Should have thrown an exception");
-		}
-		catch (Exception e)
-		{
-			assertTrue( e instanceof ClassWriterException );
-		}
+		});
 	}
-	
-	public void testSetClassName()
-	{
-		writer.setClassName( "com.company.BusinessFactory" );
-		assertEquals( "com.company", writer.getDescriptor().getPackageName() );
-		assertEquals( "BusinessFactory", writer.getDescriptor().getClassName() );
+
+	@Test
+	public void testSetClassName() {
+		writer.setClassName("com.company.BusinessFactory");
+		assertEquals("com.company", writer.getDescriptor().getPackageName());
+		assertEquals("BusinessFactory", writer.getDescriptor().getClassName());
 	}
-	
+
 }
