@@ -19,42 +19,35 @@ import java.util.Calendar;
 import java.util.Date;
 import java.util.GregorianCalendar;
 
+import org.junit.jupiter.api.Test;
+
 import com.advancedpwr.samples.Person;
 
-public class DateCalendarTest extends AbstractRecorderTest
-{
+public class DateCalendarTest extends AbstractRecorderTest {
 
-	
-	protected void setUp()
-	{
-		setWriteFiles();
-		super.setUp();
-	}
-
-	public void testRecordPerson() throws Exception
-	{
+	@Test
+	public void testRecordPerson() throws Exception {
 		Person person = Person.createExamplePerson();
 		Person dad = person.getDad();
 		Calendar calendar = GregorianCalendar.getInstance();
-		calendar.setTimeInMillis( 123456789 );
-		dad.setBirthday( calendar );
-		dad.setAnniversary( new Date( 987654321 ) );
-		recorder.setClassName( "com.example.date.PersonFactory1" );
-		recorder.record( person );
+		calendar.setTimeInMillis(123456789);
+		dad.setBirthday(calendar);
+		dad.setAnniversary(new Date(987654321));
+		recorder.setClassName("com.example.date.PersonFactory1");
+		recorder.record(person);
 		assertResult();
 	}
 
-	
-	public void testRecordPerson_sql_date() throws Exception
-	{
+	@Test
+	public void testRecordPerson_sql_date() throws Exception {
 		Person person = Person.createExamplePerson();
 		Person dad = person.getDad();
 		Calendar calendar = GregorianCalendar.getInstance();
-		calendar.setTimeInMillis( 123456789 );
-		dad.setBirthday( calendar );
-		dad.setAnniversary( new java.sql.Date( 987654321 ) );
-		recorder.setClassName( "com.example.date.PersonFactory2" );
-		recorder.record( person );
+		calendar.setTimeInMillis(123456789);
+		dad.setBirthday(calendar);
+		dad.setAnniversary(new java.sql.Date(987654321));
+		recorder.setClassName("com.example.date.PersonFactory2");
+		recorder.record(person);
 		assertResult();
 	}
 }
