@@ -15,73 +15,65 @@
  */
 package com.advancedpwr.record;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.fail;
+
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Vector;
 
+import org.junit.jupiter.api.Test;
+
 import com.advancedpwr.samples.ListExample;
 
-public class ListRecorderTest extends AbstractRecorderTest
-{
+public class ListRecorderTest extends AbstractRecorderTest {
+	@Test
+	public void testWriteObject_list() {
+		Vector<String> vector = new Vector<String>();
+		vector.add("entry 1");
+		vector.add("entry 2");
 
-	@Override
-	protected void setUp()
-	{
-		// TODO Auto-generated method stub
-		super.setUp();
-	}
-
-	public void testWriteObject_list()
-	{
-		Vector vector = new Vector();
-		vector.add( "entry 1" );
-		vector.add( "entry 2" );
-		
 		ListExample listHolder = new ListExample();
-		listHolder.setList( vector );
-		recorder.setClassName( "com.example.list.ListExampleVectorFactory" );
-		recorder.record( listHolder );
+		listHolder.setList(vector);
+		recorder.setClassName("com.example.list.ListExampleVectorFactory");
+		recorder.record(listHolder);
 		assertResult();
 	}
-	
-	public void testSubstitute()
-	{
-		recorder.useTypeForInterface( List.class, ArrayList.class );
-		
-		MyList vector = new MyList();
-		vector.add( "entry 1" );
-		vector.add( "entry 2" );
-		
+
+	@Test
+	public void testSubstitute() {
+		recorder.useTypeForInterface(List.class, ArrayList.class);
+
+		MyList<String> vector = new MyList<String>();
+		vector.add("entry 1");
+		vector.add("entry 2");
+
 		ListExample listHolder = new ListExample();
-		listHolder.setList( vector );
-		recorder.setClassName( "com.example.list.ListExampleSubstitutionFactory" );
-		recorder.useTypeForInterface( List.class, ArrayList.class );
-		recorder.record( listHolder );
-		
+		listHolder.setList(vector);
+		recorder.setClassName("com.example.list.ListExampleSubstitutionFactory");
+		recorder.useTypeForInterface(List.class, ArrayList.class);
+		recorder.record(listHolder);
+
 		assertResult();
 	}
-	
-	public void testSubstitute_failure()
-	{
-		recorder.useTypeForInterface( List.class, ArrayList.class );
-		
-		Vector vector = new Vector();
-		vector.add( "entry 1" );
-		vector.add( "entry 2" );
-		
+
+	public void testSubstitute_failure() {
+		recorder.useTypeForInterface(List.class, ArrayList.class);
+
+		Vector<String> vector = new Vector<String>();
+		vector.add("entry 1");
+		vector.add("entry 2");
+
 		ListExample listHolder = new ListExample();
-		listHolder.setList( vector );
-		
-		try
-		{
-			recorder.useTypeForInterface( List.class, HashMap.class );
-			fail( "Should have thrown an exception" );
+		listHolder.setList(vector);
+
+		try {
+			recorder.useTypeForInterface(List.class, HashMap.class);
+			fail("Should have thrown an exception");
+		} catch (Exception e) {
+			assertEquals("class java.util.HashMap cannot be used for interface java.util.List", e.getMessage());
 		}
-		catch ( Exception e )
-		{
-			assertEquals( "class java.util.HashMap cannot be used for interface java.util.List", e.getMessage() );
-		}
-		
+
 	}
 }
