@@ -1,31 +1,29 @@
 package com.advancedpwr.record;
 
+import org.junit.jupiter.api.Test;
+
 import com.advancedpwr.record.factory.MockBeanFactory;
 import com.advancedpwr.samples.Person;
 
-public class MockBeanRecorderTest extends AbstractRecorderTest
-{
-	
-	protected void setUp()
-	{
+public class MockBeanRecorderTest extends AbstractRecorderTest {
+
+	public MockBeanRecorderTest() {
 		setWriteFiles();
 		recorder = new MockBeanRecorder();
-		configureRecorder( recorder );
+		configureRecorder(recorder);
 	}
 
-	protected MockBeanRecorder getRecorder()
-	{
-		return (MockBeanRecorder)recorder;
-	}
-	public void testMockBean()
-	{
-		recorder.setClassName( "com.example.mock.MockBeanPersonFactory" );
-		recorder.setSuperClass( MockBeanFactory.class );
-		getRecorder().mock( Person.class );
-		
+	protected MockBeanRecorder getRecorder() { return (MockBeanRecorder) recorder; }
+
+	@Test
+	public void testMockBean() {
+		recorder.setClassName("com.example.mock.MockBeanPersonFactory");
+		recorder.setSuperClass(MockBeanFactory.class);
+		getRecorder().mock(Person.class);
+
 		Person person = Person.createExamplePerson();
-		recorder.record( person );
-		
+		recorder.record(person);
+
 		assertResult();
 	}
 }
