@@ -2,7 +2,8 @@ package com.advancedpwr.record;
 
 import java.util.Vector;
 
-public class MyList extends Vector
-{
+public class MyList<T> extends Vector<T> {
+
+	private static final long serialVersionUID = 1L;
 
 }
