@@ -1,21 +1,20 @@
 package com.advancedpwr.record.factory;
 
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+
 import java.util.Locale;
 
-import junit.framework.TestCase;
+import org.junit.jupiter.api.Test;
 
-public class MockBeanFactoryTest extends TestCase
-{
+public class MockBeanFactoryTest {
 
-	
-	public void testNewInstance()
-	{
-		MockBeanFactory factory = new MockBeanFactory()
-		{
+	@Test
+	public void testNewInstance() {
+		MockBeanFactory factory = new MockBeanFactory() {
 		};
-		
-		Locale locale = (Locale)factory.newInstance( Locale.class );
-		assertNotNull( locale );
+
+		Locale locale = (Locale) factory.newInstance(Locale.class);
+		assertNotNull(locale);
 	}
 
 }
