@@ -2,23 +2,13 @@ package com.advancedpwr.record.inspect;
 
 import com.advancedpwr.samples.Person;
 
-public class Ship
-{
+public class Ship {
 	protected Person fieldCaptain;
 
-	public Person getCaptain()
-	{
-		return fieldCaptain;
-	}
+	public Person getCaptain() { return fieldCaptain; }
 
-	public void setCaptain( Person captain )
-	{
-		fieldCaptain = captain;
-	}
-	
-	public String getCaptainsName()
-	{
-		return getCaptain().getName();
-	}
-	
+	public void setCaptain(Person captain) { fieldCaptain = captain; }
+
+	public String getCaptainsName() { return getCaptain().getName(); }
+
 }
